requirements.txt
langchain
streamlit
python-dotenv
langchain-community 
langchain-groq
langgraph
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
langchain-google-community
